Remove non-portable source commands
`source(1)` is a bashism and is equivalent to the portable `.(1)`, but
as each commond is run in a new shell, spawning a shell to source a file
and exit is at best noop.

Some SUT used by acceptance tests do not use bash(1) as the default
shell, which cause CI failures because `source(1)` is not a valid
command.

Because we set variables in profile files, assume they are sourced
during shell startup will be available to next spawned shells.